Organize events handler code
