+++ Blockzilla/Onboarding/OnboardingConstants.swift
@@ -1,0 +1,13 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import Foundation
+
+struct OnboardingConstants {
+    static let prefIntroDone = "IntroDone"
+    static let prefIntroVersion = 2
+    static let prefWhatsNewDone = "WhatsNewDone"
+    static let prefWhatsNewCounter = "WhatsNewCounter"
+    
+}
